requirements.txt
aiohttp==3.7.4.post0
appdirs==1.4.4
async-timeout==3.0.1
attrs==21.2.0
black==21.8b0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
cffi==1.14.6
chardet==4.0.0
click==8.0.1
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
defer==1.0.6
discord.py==1.7.3
distro==1.4.0
distro-info==0.23ubuntu1
entrypoints==0.3
flake8==3.9.2
hidpidaemon==18.4.6
httplib2==0.14.0
idna==3.2
iniconfig==1.1.1
kernelstub==3.1.4
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
louis==3.12.0
macaroonbakery==1.3.1
mccabe==0.6.1
multidict==5.1.0
mypy-extensions==0.4.3
netifaces==0.10.4
oauthlib==3.1.0
packaging==20.9
pathspec==0.9.0
pip==21.2.3
platformdirs==2.3.0
pluggy==0.13.1
pop-transition==1.1.2
protobuf==3.6.1
pubsub==0.1.2
py==1.10.0
pycairo==1.16.2
pycodestyle==2.7.0
pycparser==2.20
pycups==1.9.73
pydbus==0.6.0
pyflakes==2.3.1
PyGObject==3.36.0
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==6.2.3
python-apt==2.1.2pop0-1587756471-20.04-cd2988e
python-debian==0.1.36ubuntu1
python-xlib==0.23
pytz==2019.3
pyxdg==0.26
PyYAML==5.3.1
regex==2021.4.4
repoman==1.2.2
requests==2.22.0
requests-unixsocket==0.2.0
SecretStorage==2.3.1
setuptools==45.2.0
simplejson==3.16.0
six==1.14.0
systemd-python==234
toml==0.10.2
tomli==1.2.1
typed-ast==1.4.3
typing-extensions==3.10.0.2
ubuntu-advantage-tools==27.2
ufw==0.36
urllib3==1.25.8
wadllib==1.3.3
wheel==0.34.2
xmltodict==0.12.0
yarl==1.6.3
youtube-dl==2021.6.6
